update function added

+++ JSONGenwithpostres/hz_gen/hz_gen.go
@@ -96,33 +96,21 @@ func ClearGateway() {
 		log.Fatal("Failed to retrieve current working directory:", err)
 	}
 
+	// Delete directories recursively
 	err = filepath.Walk(wd, func(path string, info os.FileInfo, err error) error {
 		if err != nil {
-			fmt.Print("hello")
+			fmt.Println("Error occurred:", err)
 			return err
 		}
 
-		if info.IsDir() {
-			// Delete directories
-			if path != wd {
-				err := os.RemoveAll(path)
-				if err != nil {
-					log.Println("Failed to delete directory:", path, "-", err)
-				} else {
-					log.Println("Deleted directory:", path)
-				}
+		if info.IsDir() && path != wd {
+			err := os.RemoveAll(path)
+			if err != nil {
+				log.Println("Failed to delete directory:", path, "-", err)
+			} else {
+				log.Println("Deleted directory:", path)
 			}
-		} else {
-			// Delete files
-			filename := info.Name()
-			if !strings.HasPrefix(filename, "gateway") {
-				err := os.Remove(path)
-				if err != nil {
-					log.Println("Failed to delete file:", path, "-", err)
-				} else {
-					log.Println("Deleted file:", path)
-				}
-			}
+			return filepath.SkipDir
 		}
 
 		return nil
@@ -130,6 +118,27 @@ func ClearGateway() {
 
 	if err != nil {
 		log.Fatal(err)
+	}
+
+	// Delete files (excluding directories)
+	files, err := ioutil.ReadDir(wd)
+	if err != nil {
+		log.Fatal("Failed to read directory:", err)
+	}
+
+	for _, file := range files {
+		if !file.IsDir() {
+			filename := file.Name()
+			if !strings.HasPrefix(filename, "gateway") {
+				filePath := filepath.Join(wd, filename)
+				err := os.Remove(filePath)
+				if err != nil {
+					log.Println("Failed to delete file:", filePath, "-", err)
+				} else {
+					log.Println("Deleted file:", filePath)
+				}
+			}
+		}
 	}
 
 	log.Println("Deletion completed.")
